Function libsignal_service::attachment_cipher::encrypt_in_place

pub fn encrypt_in_place(iv: [u8; 16], key: [u8; 64], plaintext: &mut Vec<u8>)

Encrypts an attachment in place, given the key material.

The Vec will be reused when it has enough space to house the MAC, otherwise reallocation might happen.