libsignal_service

Module prelude

Re-exports

pub use crate::cipher::ServiceCipher;

pub use crate::configuration::ServiceConfiguration;

pub use crate::configuration::ServiceCredentials;

pub use crate::configuration::SignalingKey;

pub use crate::content::Content;

pub use crate::envelope::Envelope;

pub use crate::groups_v2::AccessControl;

pub use crate::groups_v2::Group;

pub use crate::groups_v2::Member;

pub use crate::groups_v2::PendingMember;

pub use crate::groups_v2::RequestingMember;

pub use crate::groups_v2::Timer;

pub use crate::master_key::MasterKey;

pub use crate::master_key::MasterKeyStore;

pub use crate::master_key::StorageServiceKey;

pub use crate::proto::attachment_pointer::AttachmentIdentifier;

pub use crate::proto::sync_message::Contacts;

pub use crate::proto::AttachmentPointer;

pub use crate::push_service::PushService;

pub use crate::push_service::ServiceError;

pub use crate::receiver::MessageReceiver;

pub use crate::sender::MessageSender;

pub use crate::sender::MessageSenderError;

pub use crate::session_store::SessionStoreExt;

pub use phonenumber;

Structs

DeviceId

The type used in memory to represent a device, i.e. a particular Signal client instance which represents some user.

GroupMasterKey

GroupSecretParams

InvalidDeviceId

Error for trying to construct a DeviceId with an invalid value.

ProfileKey

Uuid

A Universally Unique Identifier (UUID).

UuidError

A general error that can occur when working with UUIDs.

Traits

IdentityKeyStore

Interface defining the identity store, which may be in-memory, on-disk, etc.

ProtobufMessage

A Protocol Buffers message.

Derive Macros

ProtobufMessage