Module libsignal_service::prelude::protocol

Structs

IdentityKey

IdentityKeyPair

KeyPair

PreKeyRecord

PrivateKey

ProtocolAddress

Represents a unique Signal client instance as (<user ID>, <device ID>) pair.

PublicKey

SessionRecord

SignedPreKeyRecord

Enums

Direction

SignalProtocolError

Traits

IdentityKeyStore

PreKeyStore

SessionStore

SessionStoreExt

This is additional functions required to handle session deletion. It might be a candidate for inclusion into the bigger SessionStore trait.

SignedPreKeyStore

Type Definitions

Context