Trait libsignal_service::prelude::protocol::SessionStoreExt

pub trait SessionStoreExt: SessionStore {
    fn get_sub_device_sessions<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u32>, SignalProtocolError>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn delete_session<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        address: &'life1 ProtocolAddress
    ) -> Pin<Box<dyn Future<Output = Result<(), SignalProtocolError>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn delete_all_sessions<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        address: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<usize, SignalProtocolError>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn delete_service_addr_device_session<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        address: &'life1 ServiceAddress, 
        device_id: u32
    ) -> Pin<Box<dyn Future<Output = Result<usize, SignalProtocolError>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

This is additional functions required to handle session deletion. It might be a candidate for inclusion into the bigger SessionStore trait.

Required methods

fn get_sub_device_sessions<'life0, 'life1, 'async_trait>(
&'life0 self,
name: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Vec<u32>, SignalProtocolError>> + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get the IDs of all known devices with active sessions for a recipient.

fn delete_session<'life0, 'life1, 'async_trait>(
&'life0 self,
address: &'life1 ProtocolAddress
) -> Pin<Box<dyn Future<Output = Result<(), SignalProtocolError>> + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Remove a session record for a recipient ID + device ID tuple.

fn delete_all_sessions<'life0, 'life1, 'async_trait>(
&'life0 self,
address: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<usize, SignalProtocolError>> + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Remove the session records corresponding to all devices of a recipient ID.

Returns the number of deleted sessions.

Provided methods

fn delete_service_addr_device_session<'life0, 'life1, 'async_trait>(
&'life0 self,
address: &'life1 ServiceAddress,
device_id: u32
) -> Pin<Box<dyn Future<Output = Result<usize, SignalProtocolError>> + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Remove a session record for a recipient ID + device ID tuple.