Module libsignal_service::proto

Modules

access_control
Nested message and enum types in AccessControl.

attachment_pointer
Nested message and enum types in AttachmentPointer.

call_message
Nested message and enum types in CallMessage.

contact_details
Nested message and enum types in ContactDetails.

data_message
Nested message and enum types in DataMessage.

envelope
Nested message and enum types in Envelope.

group_attribute_blob
Nested message and enum types in GroupAttributeBlob.

group_change
Nested message and enum types in GroupChange.

group_changes
Nested message and enum types in GroupChanges.

group_context
Nested message and enum types in GroupContext.

group_details
Nested message and enum types in GroupDetails.

group_invite_link
Nested message and enum types in GroupInviteLink.

member
Nested message and enum types in Member.

pack
Nested message and enum types in Pack.

receipt_message
Nested message and enum types in ReceiptMessage.

sender_certificate
Nested message and enum types in SenderCertificate.

server_certificate
Nested message and enum types in ServerCertificate.

sync_message
Nested message and enum types in SyncMessage.

typing_message
Nested message and enum types in TypingMessage.

unidentified_sender_message
Nested message and enum types in UnidentifiedSenderMessage.

verified
Nested message and enum types in Verified.

web_socket_message
Nested message and enum types in WebSocketMessage.

Structs

AccessControl

AttachmentPointer

AvatarUploadAttributes

CallMessage

ContactDetails

Content

DataMessage

DecryptedApproveMember

DecryptedGroup
Decrypted version of message Group Keep field numbers in step

DecryptedGroupChange
Decrypted version of message GroupChange.Actions Keep field numbers in step

DecryptedGroupJoinInfo

DecryptedMember
Decrypted version of Member Keep field numbers in step

DecryptedModifyMemberRole

DecryptedPendingMember

DecryptedPendingMemberRemoval

DecryptedRequestingMember

DecryptedString

DecryptedTimer

Envelope

Group

GroupAttributeBlob

GroupChange

GroupChanges

GroupContext

GroupContextV2

GroupDetails

GroupExternalCredential

GroupInviteLink

GroupJoinInfo

Member

NullMessage

Pack

PendingMember

ProvisionEnvelope

ProvisionMessage

ProvisioningUuid

ReceiptMessage

RequestingMember

SenderCertificate

ServerCertificate

SyncMessage

TypingMessage

UnidentifiedSenderMessage

Verified

WebSocketMessage

WebSocketRequestMessage

WebSocketResponseMessage

Enums

ProvisioningVersion