Struct libsignal_service::proto::sync_message::FetchLatest

pub struct FetchLatest {
    pub type: Option<i32>,
}

Fields
type: Option<i32>
Implementations

impl FetchLatest

pub fn type(&self) -> Type

Returns the enum value of type, or the default if the field is unset or set to an invalid enum value.

pub fn set_type(&mut self, value: Type)

Sets type to the provided enum value.