Struct libsignal_service::sender::MessageSender

pub struct MessageSender<Service, S, I, SP, P, SK, R> { /* private fields */ }

Equivalent of Java’s SignalServiceMessageSender.

Implementations

impl<Service, S, I, SP, P, SK, R> MessageSender<Service, S, I, SP, P, SK, R>where
Service: PushService + Clone,
S: SessionStore + SessionStoreExt + Sync + Clone,
I: IdentityKeyStore + Clone,
SP: SignedPreKeyStore + Clone,
SK: SenderKeyStore + Clone,
P: PreKeyStore + Clone,
R: Rng + CryptoRng + Clone,

pub fn new(
ws: SignalWebSocket,
service: Service,
cipher: ServiceCipher<S, I, SP, P, SK, R>,
csprng: R,
session_store: S,
identity_key_store: I,
local_address: ServiceAddress,
device_id: DeviceId
) -> Self

pub async fn upload_attachment(
&mut self,
spec: AttachmentSpec,
contents: Vec<u8>
) -> Result<AttachmentPointer, AttachmentUploadError>

Encrypts and uploads an attachment

Contents are accepted as an owned, plain text Vec, because encryption happens in-place.

pub async fn send_message(
&mut self,
recipient: &ServiceAddress,
unidentified_access: Option<&SenderCertificate>,
message: impl Into<ContentBody>,
timestamp: u64,
online: bool
) -> SendMessageResult

Send a message content to a single recipient.

pub async fn send_message_to_group(
&mut self,
recipients: impl AsRef<[ServiceAddress]>,
unidentified_access: Option<&SenderCertificate>,
message: impl Into<ContentBody>,
timestamp: u64,
online: bool
) -> Vec<SendMessageResult>

Send a message to the recipients in a group.

pub async fn send_groups_details<Groups>(
&mut self,
recipient: &ServiceAddress,
unidentified_access: Option<&SenderCertificate>,
groups: Groups,
online: bool
) -> Result<(), MessageSenderError>where
Groups: IntoIterator<Item = GroupDetails>,

Upload group details to the CDN and send a sync message

pub async fn send_contact_details<Contacts>(
&mut self,
recipient: &ServiceAddress,
unidentified_access: Option<&SenderCertificate>,
contacts: Contacts,
online: bool,
complete: bool
) -> Result<(), MessageSenderError>where
Contacts: IntoIterator<Item = ContactDetails>,

Upload contact details to the CDN and send a sync message