Crate libsignal_service

Re-exports

pub use libsignal_protocol as protocol;

pub use zkgroup;

pub use crate::service_address::*;

Modules

attachment_cipher

cipher

configuration

content

envelope

groups_v2

Everything needed to support Signal Groups v2

master_key

messagepipe

models

pre_keys

prelude

profile_cipher

profile_name

proto

provisioning

push_service

receiver

sender

service_address

session_store

sticker_cipher

unidentified_access

utils

websocket

Structs

AccountManager

Profile

Enums

ProfileManagerError

Constants

GROUP_LEAVE_FLAG

GROUP_LEAVE_FLAG signals that this message is a group leave message.

GROUP_UPDATE_FLAG

GROUP_UPDATE_FLAG signals that this message updates the group membership or name.

USER_AGENT

Functions

decrypt_device_name