Crate libsignal_service

Re-exports

pub use crate::service_address::*;

Modules

attachment_cipher

cipher

configuration

content

envelope

groups_v2

Everything needed to support Signal Groups v2

messagepipe

models

pre_keys

prelude

profile_cipher

profile_name

proto

provisioning

push_service

receiver

sender

service_address

utils

Structs

AccountManager

Profile

Enums

ProfileManagerError

Constants

GROUP_LEAVE_FLAG

GROUP_LEAVE_FLAG signals that this message is a group leave message.

GROUP_UPDATE_FLAG

GROUP_UPDATE_FLAG signals that this message updates the group membership or name.

USER_AGENT

Traits

MaybeSend

This trait allows for the conditional support of Send compatible futures depending on whether or not the unsend-futures feature flag is enabled. As this feature is disabled by default, Send is supported by default.