Trait libsignal_service::groups_v2::CredentialsCache

pub trait CredentialsCache {
    fn clear(&mut self) -> Result<(), CredentialsCacheError>;
    fn get(
        &self,
        key: &i64
    ) -> Result<Option<&AuthCredentialResponse>, CredentialsCacheError>;
    fn write(
        &mut self,
        map: HashMap<i64, AuthCredentialResponse>
    ) -> Result<(), CredentialsCacheError>;
}

Global cache for groups v2 credentials, as demonstrated in the libsignal-service java library of Signal-Android.

A basic in-memory implementation is provided with InMemoryCredentialsCache.

Required Methods

fn clear(&mut self) -> Result<(), CredentialsCacheError>

fn get(
&self,
key: &i64
) -> Result<Option<&AuthCredentialResponse>, CredentialsCacheError>

Get an entry of the cache, key usually represents the day number since EPOCH.

fn write(
&mut self,
map: HashMap<i64, AuthCredentialResponse>
) -> Result<(), CredentialsCacheError>

Overwrite the entire contents of the cache with new data.