Module libsignal_service::prelude::protocol

Structs

DeviceId

The type used in memory to represent a device, i.e. a particular Signal client instance which represents some user.

IdentityKey

A public key that represents the identity of a user.

IdentityKeyPair

The private identity of a user.

KeyPair

PreKeyId

A unique identifier selecting among this client’s known pre-keys.

PreKeyRecord

PrivateKey

ProtocolAddress

Represents a unique Signal client instance as (<user ID>, <device ID>) pair.

PublicKey

SenderCertificate

SenderKeyRecord

SessionRecord

SignedPreKeyId

A unique identifier selecting among this client’s known signed pre-keys.

SignedPreKeyRecord

Enums

Direction

Each Signal message can be considered to have exactly two participants, a sender and receiver.

SignalProtocolError

Traits

IdentityKeyStore

Interface defining the identity store, which may be in-memory, on-disk, etc.

PreKeyStore

Interface for storing pre-keys downloaded from a server.

ProtocolStore

Mixes in all the store interfaces defined in this module.

SenderKeyStore

Interface for storing sender key records, allowing multiple keys per user.

SessionStore

Interface for a Signal client instance to store a session associated with another particular separate Signal client instance.

SessionStoreExt

This is additional functions required to handle session deletion. It might be a candidate for inclusion into the bigger SessionStore trait.

SignedPreKeyStore

Interface for storing signed pre-keys downloaded from a server.

Type Definitions

Context

Handle to FFI-provided context object.