Module libsignal_service::prelude::protocol

Structs

DeviceId

The type used in memory to represent a device, i.e. a particular Signal client instance which represents some user.

IdentityKey

A public key that represents the identity of a user.

IdentityKeyPair

The private identity of a user.

KeyPair

PreKeyId

A unique identifier selecting among this client’s known pre-keys.

PreKeyRecord

PrivateKey

ProtocolAddress

Represents a unique Signal client instance as (<user ID>, <device ID>) pair.

PublicKey

SenderCertificate

SenderKeyRecord

SessionRecord

SignedPreKeyId

A unique identifier selecting among this client’s known signed pre-keys.

SignedPreKeyRecord

Enums

Direction

SignalProtocolError

Traits

IdentityKeyStore

PreKeyStore

SenderKeyStore

SessionStore

SessionStoreExt

This is additional functions required to handle session deletion. It might be a candidate for inclusion into the bigger SessionStore trait.

SignedPreKeyStore

Type Definitions

Context