Trait libsignal_service::prelude::ProtobufMessage

pub trait ProtobufMessage: Send + Sync + Debug {
    pub fn encoded_len(&self) -> usize;

    pub fn clear(&mut self);

    pub fn encode<B>(&self, buf: &mut B) -> Result<(), EncodeError>
    where
        B: BufMut,
    { ... }

    pub fn encode_length_delimited<B>(
        &self, 
        buf: &mut B
    ) -> Result<(), EncodeError>
    where
        B: BufMut,
    { ... }

    pub fn decode<B>(buf: B) -> Result<Self, DecodeError>
    where
        Self: Default,
        B: Buf,
    { ... }

    pub fn decode_length_delimited<B>(buf: B) -> Result<Self, DecodeError>
    where
        Self: Default,
        B: Buf,
    { ... }

    pub fn merge<B>(&mut self, buf: B) -> Result<(), DecodeError>
    where
        B: Buf,
    { ... }

    pub fn merge_length_delimited<B>(
        &mut self, 
        buf: B
    ) -> Result<(), DecodeError>
    where
        B: Buf,
    { ... }
}

A Protocol Buffers message.

Required methods

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.

pub fn clear(&mut self)

Clears the message, resetting all fields to their default.

Loading content...
Provided methods

pub fn encode<B>(&self, buf: &mut B) -> Result<(), EncodeError> where
B: BufMut,

Encodes the message to a buffer.

An error will be returned if the buffer does not have sufficient capacity.

pub fn encode_length_delimited<B>(&self, buf: &mut B) -> Result<(), EncodeError> where
B: BufMut,

Encodes the message with a length-delimiter to a buffer.

An error will be returned if the buffer does not have sufficient capacity.

pub fn decode<B>(buf: B) -> Result<Self, DecodeError> where
Self: Default,
B: Buf,

Decodes an instance of the message from a buffer.

The entire buffer will be consumed.

pub fn decode_length_delimited<B>(buf: B) -> Result<Self, DecodeError> where
Self: Default,
B: Buf,

Decodes a length-delimited instance of the message from the buffer.

pub fn merge<B>(&mut self, buf: B) -> Result<(), DecodeError> where
B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

The entire buffer will be consumed.

pub fn merge_length_delimited<B>(&mut self, buf: B) -> Result<(), DecodeError> where
B: Buf,

Decodes a length-delimited instance of the message from buffer, and merges it into self.

Loading content...