Trait libsignal_service::prelude::ProtobufMessage

pub trait ProtobufMessage: Debug + Send + Sync {
    fn encoded_len(&self) -> usize;
    fn clear(&mut self);

    fn encode<B>(&self, buf: &mut B) -> Result<(), EncodeError>
    where
        B: BufMut,
    { ... }
    fn encode_to_vec(&self) -> Vec<u8, Global> { ... }
    fn encode_length_delimited<B>(&self, buf: &mut B) -> Result<(), EncodeError>
    where
        B: BufMut,
    { ... }
    fn encode_length_delimited_to_vec(&self) -> Vec<u8, Global> { ... }
    fn decode<B>(buf: B) -> Result<Self, DecodeError>
    where
        B: Buf,
        Self: Default,
    { ... }
    fn decode_length_delimited<B>(buf: B) -> Result<Self, DecodeError>
    where
        B: Buf,
        Self: Default,
    { ... }
    fn merge<B>(&mut self, buf: B) -> Result<(), DecodeError>
    where
        B: Buf,
    { ... }
    fn merge_length_delimited<B>(&mut self, buf: B) -> Result<(), DecodeError>
    where
        B: Buf,
    { ... }
}

A Protocol Buffers message.

Required Methods

fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.

fn clear(&mut self)

Clears the message, resetting all fields to their default.

Provided Methods

fn encode<B>(&self, buf: &mut B) -> Result<(), EncodeError> where
B: BufMut,

Encodes the message to a buffer.

An error will be returned if the buffer does not have sufficient capacity.

fn encode_to_vec(&self) -> Vec<u8, Global>

Encodes the message to a newly allocated buffer.

fn encode_length_delimited<B>(&self, buf: &mut B) -> Result<(), EncodeError> where
B: BufMut,

Encodes the message with a length-delimiter to a buffer.

An error will be returned if the buffer does not have sufficient capacity.

fn encode_length_delimited_to_vec(&self) -> Vec<u8, Global>

Encodes the message with a length-delimiter to a newly allocated buffer.

fn decode<B>(buf: B) -> Result<Self, DecodeError> where
B: Buf,
Self: Default,

Decodes an instance of the message from a buffer.

The entire buffer will be consumed.

fn decode_length_delimited<B>(buf: B) -> Result<Self, DecodeError> where
B: Buf,
Self: Default,

Decodes a length-delimited instance of the message from the buffer.

fn merge<B>(&mut self, buf: B) -> Result<(), DecodeError> where
B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

The entire buffer will be consumed.

fn merge_length_delimited<B>(&mut self, buf: B) -> Result<(), DecodeError> where
B: Buf,

Decodes a length-delimited instance of the message from buffer, and merges it into self.