Trait libsignal_service::push_service::PushService

pub trait PushService: MaybeSend {
    type WebSocket: WebSocketService;
    type ByteStream: AsyncRead + Unpin;

    fn get_json<'life0, 'life1, 'async_trait, T>(
        &'life0 mut self, 
        service: Endpoint, 
        path: &'life1 str, 
        credentials_override: HttpAuthOverride
    ) -> Pin<Box<dyn Future<Output = Result<T, ServiceError>> + Send + 'async_trait>>
    where
        for<'de> T: Deserialize<'de>,
        T: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn delete_json<'life0, 'life1, 'async_trait, T>(
        &'life0 mut self, 
        service: Endpoint, 
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<T, ServiceError>> + Send + 'async_trait>>
    where
        for<'de> T: Deserialize<'de>,
        T: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn put_json<'life0, 'life1, 'async_trait, D, S>(
        &'life0 mut self, 
        service: Endpoint, 
        path: &'life1 str, 
        credentials_override: HttpAuthOverride, 
        value: S
    ) -> Pin<Box<dyn Future<Output = Result<D, ServiceError>> + Send + 'async_trait>>
    where
        for<'de> D: Deserialize<'de>,
        S: MaybeSend + Serialize,
        D: 'async_trait,
        S: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_protobuf<'life0, 'life1, 'async_trait, T>(
        &'life0 mut self, 
        service: Endpoint, 
        path: &'life1 str, 
        credentials_override: HttpAuthOverride
    ) -> Pin<Box<dyn Future<Output = Result<T, ServiceError>> + Send + 'async_trait>>
    where
        T: Default + ProtobufMessage,
        T: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn put_protobuf<'life0, 'life1, 'async_trait, D, S>(
        &'life0 mut self, 
        service: Endpoint, 
        path: &'life1 str, 
        value: S
    ) -> Pin<Box<dyn Future<Output = Result<D, ServiceError>> + Send + 'async_trait>>
    where
        D: Default + ProtobufMessage,
        S: Sized + ProtobufMessage,
        D: 'async_trait,
        S: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_from_cdn<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        cdn_id: u32, 
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Self::ByteStream, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn post_to_cdn0<'s, 'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait, C>(
        &'life0 mut self, 
        path: &'life1 str, 
        value: &'life2 [(&'life3 str, &'life4 str)], 
        file: Option<(&'life5 str, &'s mut C)>
    ) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>>
    where
        's: 'async_trait,
        C: 'async_trait + Read + Send + 's,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        'life5: 'async_trait,
        Self: 'async_trait;
    fn ws<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        path: &'life1 str, 
        credentials: Option<ServiceCredentials>
    ) -> Pin<Box<dyn Future<Output = Result<(Self::WebSocket, <Self::WebSocket as WebSocketService>::Stream), ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn devices<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<DeviceInfo>, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn unlink_device<'life0, 'async_trait>(
        &'life0 mut self, 
        id: i64
    ) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_pre_key_status<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<PreKeyStatus, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn register_pre_keys<'life0, 'async_trait>(
        &'life0 mut self, 
        pre_key_state: PreKeyState
    ) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_attachment_by_id<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        id: &'life1 str, 
        cdn_id: u32
    ) -> Pin<Box<dyn Future<Output = Result<Self::ByteStream, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_attachment<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        ptr: &'life1 AttachmentPointer
    ) -> Pin<Box<dyn Future<Output = Result<Self::ByteStream, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn send_messages<'a, 'life0, 'async_trait>(
        &'life0 mut self, 
        messages: OutgoingPushMessages<'a>
    ) -> Pin<Box<dyn Future<Output = Result<SendMessageResponse, ServiceError>> + Send + 'async_trait>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_attachment_v2_upload_attributes<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<AttachmentV2UploadAttributes, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn upload_attachment<'s, 'life0, 'life1, 'async_trait, C>(
        &'life0 mut self, 
        attrs: &'life1 AttachmentV2UploadAttributes, 
        content: &'s mut C
    ) -> Pin<Box<dyn Future<Output = Result<(u64, Vec<u8>), ServiceError>> + Send + 'async_trait>>
    where
        's: 'async_trait,
        C: 'async_trait + Read + Send + 's,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_messages<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<EnvelopeEntity>, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn whoami<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<WhoAmIResponse, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn retrieve_profile_by_id<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<SignalServiceProfile, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_pre_key<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        destination: &'life1 ServiceAddress, 
        device_id: u32
    ) -> Pin<Box<dyn Future<Output = Result<PreKeyBundle, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_pre_keys<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        destination: &'life1 ServiceAddress, 
        device_id: u32
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PreKeyBundle>, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_group<'life0, 'async_trait>(
        &'life0 mut self, 
        credentials: HttpAuth
    ) -> Pin<Box<dyn Future<Output = Result<Group, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn set_account_attributes<'life0, 'async_trait>(
        &'life0 mut self, 
        attributes: AccountAttributes
    ) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn write_profile<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
        &'life0 mut self, 
        version: &'life1 ProfileKeyVersion, 
        name: &'life2 [u8], 
        about: &'life3 [u8], 
        emoji: &'life4 [u8], 
        commitment: &'life5 ProfileKeyCommitment, 
        avatar: Option<()>
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, ServiceError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        'life5: 'async_trait,
        Self: Send + 'async_trait,
    { ... }

}

Required Associated Types

type WebSocket: WebSocketService

type ByteStream: AsyncRead + Unpin

Required Methods

fn get_json<'life0, 'life1, 'async_trait, T>(
&'life0 mut self,
service: Endpoint,
path: &'life1 str,
credentials_override: HttpAuthOverride
) -> Pin<Box<dyn Future<Output = Result<T, ServiceError>> + Send + 'async_trait>> where
for<'de> T: Deserialize<'de>,
T: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn delete_json<'life0, 'life1, 'async_trait, T>(
&'life0 mut self,
service: Endpoint,
path: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<T, ServiceError>> + Send + 'async_trait>> where
for<'de> T: Deserialize<'de>,
T: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn put_json<'life0, 'life1, 'async_trait, D, S>(
&'life0 mut self,
service: Endpoint,
path: &'life1 str,
credentials_override: HttpAuthOverride,
value: S
) -> Pin<Box<dyn Future<Output = Result<D, ServiceError>> + Send + 'async_trait>> where
for<'de> D: Deserialize<'de>,
S: MaybeSend + Serialize,
D: 'async_trait,
S: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn get_protobuf<'life0, 'life1, 'async_trait, T>(
&'life0 mut self,
service: Endpoint,
path: &'life1 str,
credentials_override: HttpAuthOverride
) -> Pin<Box<dyn Future<Output = Result<T, ServiceError>> + Send + 'async_trait>> where
T: Default + ProtobufMessage,
T: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn put_protobuf<'life0, 'life1, 'async_trait, D, S>(
&'life0 mut self,
service: Endpoint,
path: &'life1 str,
value: S
) -> Pin<Box<dyn Future<Output = Result<D, ServiceError>> + Send + 'async_trait>> where
D: Default + ProtobufMessage,
S: Sized + ProtobufMessage,
D: 'async_trait,
S: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn get_from_cdn<'life0, 'life1, 'async_trait>(
&'life0 mut self,
cdn_id: u32,
path: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Self::ByteStream, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Downloads larger files in streaming fashion, e.g. attachments.

fn post_to_cdn0<'s, 'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait, C>(
&'life0 mut self,
path: &'life1 str,
value: &'life2 [(&'life3 str, &'life4 str)],
file: Option<(&'life5 str, &'s mut C)>
) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>> where
's: 'async_trait,
C: 'async_trait + Read + Send + 's,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
'life5: 'async_trait,
Self: 'async_trait,

Upload larger file to CDN0 in legacy fashion, e.g. for attachments.

Implementations are allowed to panic when the Read instance throws an IO-Error

fn ws<'life0, 'life1, 'async_trait>(
&'life0 mut self,
path: &'life1 str,
credentials: Option<ServiceCredentials>
) -> Pin<Box<dyn Future<Output = Result<(Self::WebSocket, <Self::WebSocket as WebSocketService>::Stream), ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Provided Methods

fn devices<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Vec<DeviceInfo>, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

Fetches a list of all devices tied to the authenticated account.

This list include the device that sends the request.

fn unlink_device<'life0, 'async_trait>(
&'life0 mut self,
id: i64
) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn get_pre_key_status<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<PreKeyStatus, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn register_pre_keys<'life0, 'async_trait>(
&'life0 mut self,
pre_key_state: PreKeyState
) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn get_attachment_by_id<'life0, 'life1, 'async_trait>(
&'life0 mut self,
id: &'life1 str,
cdn_id: u32
) -> Pin<Box<dyn Future<Output = Result<Self::ByteStream, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

fn get_attachment<'life0, 'life1, 'async_trait>(
&'life0 mut self,
ptr: &'life1 AttachmentPointer
) -> Pin<Box<dyn Future<Output = Result<Self::ByteStream, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

fn send_messages<'a, 'life0, 'async_trait>(
&'life0 mut self,
messages: OutgoingPushMessages<'a>
) -> Pin<Box<dyn Future<Output = Result<SendMessageResponse, ServiceError>> + Send + 'async_trait>> where
'a: 'async_trait,
'life0: 'async_trait,
Self: Send + 'async_trait,

fn get_attachment_v2_upload_attributes<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<AttachmentV2UploadAttributes, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

Request AttachmentV2UploadAttributes

Equivalent with getAttachmentV2UploadAttributes

fn upload_attachment<'s, 'life0, 'life1, 'async_trait, C>(
&'life0 mut self,
attrs: &'life1 AttachmentV2UploadAttributes,
content: &'s mut C
) -> Pin<Box<dyn Future<Output = Result<(u64, Vec<u8>), ServiceError>> + Send + 'async_trait>> where
's: 'async_trait,
C: 'async_trait + Read + Send + 's,
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

Upload attachment to CDN

Returns attachment ID and the attachment digest

fn get_messages<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Vec<EnvelopeEntity>, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn whoami<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<WhoAmIResponse, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

Method used to check our own UUID

fn retrieve_profile_by_id<'life0, 'life1, 'async_trait>(
&'life0 mut self,
id: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<SignalServiceProfile, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

fn get_pre_key<'life0, 'life1, 'async_trait>(
&'life0 mut self,
destination: &'life1 ServiceAddress,
device_id: u32
) -> Pin<Box<dyn Future<Output = Result<PreKeyBundle, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

fn get_pre_keys<'life0, 'life1, 'async_trait>(
&'life0 mut self,
destination: &'life1 ServiceAddress,
device_id: u32
) -> Pin<Box<dyn Future<Output = Result<Vec<PreKeyBundle>, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

fn get_group<'life0, 'async_trait>(
&'life0 mut self,
credentials: HttpAuth
) -> Pin<Box<dyn Future<Output = Result<Group, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn set_account_attributes<'life0, 'async_trait>(
&'life0 mut self,
attributes: AccountAttributes
) -> Pin<Box<dyn Future<Output = Result<(), ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn write_profile<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
&'life0 mut self,
version: &'life1 ProfileKeyVersion,
name: &'life2 [u8],
about: &'life3 [u8],
emoji: &'life4 [u8],
commitment: &'life5 ProfileKeyCommitment,
avatar: Option<()>
) -> Pin<Box<dyn Future<Output = Result<Option<String>, ServiceError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
'life5: 'async_trait,
Self: Send + 'async_trait,

Writes a profile and returns the avatar URL, if one was provided.

The name, about and emoji fields are encrypted with an [ProfileCipher]. See [AccountManager] for a convenience method.

Java equivalent: writeProfile